+++ plugins/destination/plugin.go
@@ -207,6 +207,7 @@ func (p *Plugin) Write(ctx context.Context, sourceSpec specs.Source, tables sche
 	if p.spec.WriteMode == specs.WriteModeOverwriteDeleteStale {
 		tablesToDelete := tables
 		if sourceSpec.Backend != specs.BackendNone {
+			tablesToDelete = make(schema.Tables, 0, len(tables))
 			for _, t := range tables {
 				if !t.IsIncremental {
 					tablesToDelete = append(tablesToDelete, t)
@@ -236,18 +237,18 @@ func (p *Plugin) BatchingWriter() BatchingWriter {
 
 func checkDestinationColumns(tables schema.Tables) error {
 	for _, table := range tables {
-		if table.Columns.Index(schema.CqSourceNameField.Name) == -1 {
-			return fmt.Errorf("table %s is missing column %s. please consider upgrading source plugin", table.Name, schema.CqSourceNameField.Name)
+		if table.Columns.Index(schema.CqSourceNameColumn.Name) == -1 {
+			return fmt.Errorf("table %s is missing column %s. please consider upgrading source plugin", table.Name, schema.CqSourceNameColumn.Name)
 		}
 		if table.Columns.Index(schema.CqSyncTimeColumn.Name) == -1 {
-			return fmt.Errorf("table %s is missing column %s. please consider upgrading source plugin", table.Name, schema.CqSourceNameField.Name)
+			return fmt.Errorf("table %s is missing column %s. please consider upgrading source plugin", table.Name, schema.CqSourceNameColumn.Name)
 		}
-		field := table.Columns.Get(schema.CqIDColumn.Name)
-		if field != nil {
-			if !field.CreationOptions.NotNull {
+		column := table.Columns.Get(schema.CqIDColumn.Name)
+		if column != nil {
+			if !column.NotNull {
 				return fmt.Errorf("column %s.%s cannot be nullable. please consider upgrading source plugin", table.Name, schema.CqIDColumn.Name)
 			}
-			if !field.CreationOptions.Unique {
+			if !column.Unique {
 				return fmt.Errorf("column %s.%s must be unique. please consider upgrading source plugin", table.Name, schema.CqIDColumn.Name)
 			}
 		}