+++ scalar/type_test.go
@@ -1,0 +1,8 @@
+package scalar
+
+// Test for renamed types
+type _string string
+type _bool bool
+type _int8 int8
+
+type _byteSlice []byte