+++ schema/meta.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/apache/arrow/go/v13/arrow"
+	"github.com/cloudquery/plugin-sdk/v3/scalar"
 	"github.com/cloudquery/plugin-sdk/v3/types"
-	"github.com/google/uuid"
 )
 
 type ClientMeta interface {
@@ -17,11 +17,10 @@ var CqIDColumn = Column{
 	Name:        "_cq_id",
 	Type:        types.ExtensionTypes.UUID,
 	Description: "Internal CQ ID of the row",
-	CreationOptions: ColumnCreationOptions{
-		NotNull: true,
-		Unique:  true,
-	},
+	NotNull:     true,
+	Unique:      true,
 }
+
 var CqParentIDColumn = Column{
 	Name:          "_cq_parent_id",
 	Type:          types.ExtensionTypes.UUID,
@@ -36,26 +35,11 @@ var CqSyncTimeColumn = Column{
 	Type:        arrow.FixedWidthTypes.Timestamp_us,
 	Description: "Internal CQ row of when sync was started (this will be the same for all rows in a single fetch)",
 }
+
 var CqSourceNameColumn = Column{
 	Name:        "_cq_source_name",
 	Type:        arrow.BinaryTypes.String,
 	Description: "Internal CQ row that references the source plugin name data was retrieved",
-}
-
-var CqIDField = arrow.Field{
-	Name: "_cq_id",
-	Type: types.ExtensionTypes.UUID,
-	Metadata: arrow.MetadataFrom(map[string]string{
-		MetadataUnique: MetadataTrue,
-	}),
-}
-var CqSyncTimeField = arrow.Field{
-	Name: "_cq_sync_time",
-	Type: arrow.FixedWidthTypes.Timestamp_us,
-}
-var CqSourceNameField = arrow.Field{
-	Name: "_cq_source_name",
-	Type: arrow.BinaryTypes.String,
 }
 
 func parentCqUUIDResolver() ColumnResolver {
@@ -67,7 +51,7 @@ func parentCqUUIDResolver() ColumnResolver {
 		if parentCqID == nil {
 			return r.Set(c.Name, nil)
 		}
-		pUUID, ok := parentCqID.(*uuid.UUID)
+		pUUID, ok := parentCqID.(*scalar.UUID)
 		if !ok {
 			return r.Set(c.Name, nil)
 		}