+++ serve/destination.go
@@ -10,10 +10,13 @@ import (
 	"syscall"
 
 	pbv0 "github.com/cloudquery/plugin-pb-go/pb/destination/v0"
+	pbv1 "github.com/cloudquery/plugin-pb-go/pb/destination/v1"
 	pbdiscoveryv0 "github.com/cloudquery/plugin-pb-go/pb/discovery/v0"
 	servers "github.com/cloudquery/plugin-sdk/v3/internal/servers/destination/v0"
+	serversv1 "github.com/cloudquery/plugin-sdk/v3/internal/servers/destination/v1"
 	discoveryServerV0 "github.com/cloudquery/plugin-sdk/v3/internal/servers/discovery/v0"
 	"github.com/cloudquery/plugin-sdk/v3/plugins/destination"
+	"github.com/cloudquery/plugin-sdk/v3/types"
 	"github.com/getsentry/sentry-go"
 	grpczerolog "github.com/grpc-ecosystem/go-grpc-middleware/providers/zerolog/v2"
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/logging"
@@ -115,8 +118,12 @@ func newCmdDestinationServe(serve *destinationServe) *cobra.Command {
 				Plugin: serve.plugin,
 				Logger: logger,
 			})
+			pbv1.RegisterDestinationServer(s, &serversv1.Server{
+				Plugin: serve.plugin,
+				Logger: logger,
+			})
 			pbdiscoveryv0.RegisterDiscoveryServer(s, &discoveryServerV0.Server{
-				Versions: []string{"v0"},
+				Versions: []string{"v0", "v1"},
 			})
 			version := serve.plugin.Version()
 
@@ -144,6 +151,16 @@ func newCmdDestinationServe(serve *destinationServe) *cobra.Command {
 					log.Error().Err(err).Msg("Error initializing sentry")
 				}
 			}
+
+			if err := types.RegisterAllExtensions(); err != nil {
+				return err
+			}
+			defer func() {
+				if err := types.UnregisterAllExtensions(); err != nil {
+					logger.Error().Err(err).Msg("Failed to unregister extensions")
+				}
+			}()
+
 			ctx := cmd.Context()
 			c := make(chan os.Signal, 1)
 			signal.Notify(c, os.Interrupt, syscall.SIGTERM)