+++ types/json.go
@@ -20,6 +20,15 @@ func NewJSONBuilder(builder *array.ExtensionBuilder) *JSONBuilder {
 	return &JSONBuilder{ExtensionBuilder: builder}
 }
 
+func (b *JSONBuilder) AppendBytes(v []byte) {
+	if v == nil {
+		b.AppendNull()
+		return
+	}
+
+	b.ExtensionBuilder.Builder.(*array.BinaryBuilder).Append(v)
+}
+
 func (b *JSONBuilder) Append(v any) {
 	if v == nil {
 		b.AppendNull()
@@ -200,8 +209,8 @@ func (*JSONType) ExtensionName() string {
 	return "json"
 }
 
-func (e *JSONType) String() string {
-	return fmt.Sprintf("extension_type<storage=%s>", e.Storage)
+func (*JSONType) String() string {
+	return "json"
 }
 
 func (e *JSONType) MarshalJSON() ([]byte, error) {